+++ printf.h
@@ -4,5 +4,5 @@
 
 int put_char(char c);
 int put_str(char *str);
-int put_nbr_base(size_t num, char *base, size_t len);
+int put_nbr_base(unsigned int num, char *base, size_t len);
 int put_pointer_adress(void *ptr);
